Handle integrations in the extension
They used to be handled in the other extensions but now that the
extension has been moved to the Quarkiverse, we need to handle it in
this very extension.

Also fixes a few issues that had not been caught testing with 3.5.1 in
passing.

The tests won't pass until we move the build to 3.6.0.